Translations for Taxonomies (#11845)
* add yarn build to sandbox.sh

* Add mobility gem and basic usage for products

* fix spree_sample::load

* Add translations for taxons, fix loading taxons with spree_sample

* Permit :locale as param

* Add mobility ransack and join translations table for taxons

* fix ransack sorting

* Downgrade mobility-ransack to 1.0.1

* Add friendly_id-mobility gem and translatable slugs

* Small code changes connected to mobility gem

* validates taxon name presence

* Add migration for copying existing data to translatable tables

* Update tables after migrating translatable data

* move common functionality to TranslatableResource

* make TranslatableResource Concern instead of parent class

* Add mobility gem and basic usage for products

* fix spree_sample::load

* Add translations for taxons, fix loading taxons with spree_sample

* Permit :locale as param

* Add mobility ransack and join translations table for taxons

* fix ransack sorting

* Downgrade mobility-ransack to 1.0.1

* Add friendly_id-mobility gem and translatable slugs

* rebase

* validates taxon name presence

* Add migration for copying existing data to translatable tables

* Update tables after migrating translatable data

* move common functionality to TranslatableResource

* make TranslatableResource Concern instead of parent class

* automatically scope to i18n for translatable resources

* update find by name to be compatible w mobility

* Fix for PG::InvalidColumnReference in products table

* add translation scope to product test conditions

* update get_taxons in product scopes for mobility compatibility

* update product find -> ordering for mobility compatibility

* update product search by name for mobility

* Clean up plugins and set default to I18n default

* mobility-ransack issue workaround: filtering variants by product attribute

* make filtering variant by product compatible w mobility pt2

* fix taxon name validation test

* Add taxonomies to translatable resources

* rewrite sample products+taxons for mobility

* add acts_as_paranoidto product translations

* check for updated translations in update webhook

* updates for product translations acts_as_paranoid

* add unique constraint for slug per locale to db

* more updates for mobility/webhooks integration

* change migration version to 6.1 for rails 6.1 compatibility

* freeze translatable_fields constants

* update product duplicator to handle translations

* revert taxonomy translation

* delete taxonomy translation migration files

* update methods checking for changes in translated values

* use mobility dirty plugin instead of manual checks for saved changes

* account for mysql rules for select statements

* add brakeman warning to ignore list

* add path to brakeman.ignore for circleci

* add warning to brakeman (spree core)

* remove translatable resource controller parent class

* refactoring and code climate fixes

* change how translation fields are passed

* move join product translation code into a method

* add taxon find spec

* refactor product 'ordered' method

* update brakeman.ignore

* remove checks for SpreeGlobalize on product searches

* Add translations for option types and option values

* method for joining option value translations, plus tests

* refactor join translation table methods

* valid create params for failing 'resource created' tests

* update brakeman.ignore

* migration for transferring option data to translations table

* property and product property translations

* make option type sample compatible w mobility

* change rails migration version to 6.1

* add translations to property models + update relevant product scope methods

* apply auto_strip to translation class, plus other test fixes

* fix broken tests due to Mobility quirks

* update filter by brand to work with mobility

* bump version to allow for integration tests

* code climate style tweaks

* Account for possibility of existing translations from globalize

* update brakeman.ignore

* update migrations for mysql compatibility

* taxonomy translations + data transfer abstraction class

* fix broken taxonomy related tests

* update brakeman ignore

* menu translations

* fix broken menu tests

* account for issue with Rails if_exists on dropping indexes

* taxonomies: rewrite first_or_create for mobility compatibility

* menus: rewrite first_or_create for mobility compatibility

* add menu item translations

* fix tests and cleanup

* make translations accessible for deleted products

* revert menu and menu item translations

* revert menu and menu item translations part 2

* revert menu and menu item translations part 3

---------

+++ core/db/migrate/20230103144439_create_option_type_translations.rb
@@ -1,7 +1,10 @@
 class CreateOptionTypeTranslations < ActiveRecord::Migration[6.1]
   def change
     if ActiveRecord::Base.connection.table_exists? 'spree_option_type_translations'
-      remove_index :spree_option_type_translations, name: "index_spree_option_type_translations_on_spree_option_type_id", if_exists: true
+      # manually check for index since Rails if_exists does not always work correctly
+      if ActiveRecord::Migration.connection.index_exists?(:spree_option_type_translations, :spree_option_type_id)
+        remove_index :spree_option_type_translations, name: "index_spree_option_type_translations_on_spree_option_type_id", if_exists: true
+      end
     else
       create_table :spree_option_type_translations do |t|
